Moved command bar out of experiments and into oufr
