--- packages/experiments/src/CommandBar.ts
@@ -1,1 +1,0 @@
-export * from './components/CommandBar';
