--- packages/experiments/src/components/CommandBar/index.ts
@@ -1,2 +1,0 @@
-export * from './CommandBar';
-export * from './CommandBar.Props';
