+++ packages/experiments/src/index.ts
@@ -1,7 +1,5 @@
-export { CommandBar } from './CommandBar';
 export { LayoutGroup } from './LayoutGroup';
 export { FolderCover } from './FolderCover';
 export { Tile } from './Tile';
 export { TilesList } from './TilesList';
-export { CommandBar as ExperimentCommandBar } from './CommandBar';
 export { ExtendedPeoplePicker, FloatingPeoplePicker } from './ExtendedPicker';
