+++ packages/office-ui-fabric-react/src/components/CommandBar/CommandBar.scss
@@ -1,302 +1,19 @@
-@import '../../common/common';
-@import '../../common/_focusBorder';
-
-// Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See LICENSE in the project root for license information.
-
-//
-// Office UI Fabric
-// --------------------------------------------------
-// Command Bar styles
-//
-// Using Padding instead of Margin for ms-CommandBarItem so that width to
-// calculate overFlow items is calculated correctly.
-
-$CommandBar-height: 40px;
-$CommandBar-mobileHeight: 48px;
-$CommandBar-iconSize: 18px;
-$CommandBar-smallLeft: 44px;
-$CommandBar-largeLeft: 48px;
-$CommandBar-xLargeLeft: 232px;
-$CommandBar-sideAreaWidthsm: 128px;
-$CommandBar-sideAreaWidthlg: 260px;
-$CommandBar-sideAreaWidthxlg: 345px;
-
-$CommandBarItem-height: $CommandBar-height;
-$CommandBarItem-marginTop: 13px;
-$CommandBarItem-textHeight: 16px;
-$CommandBarItem-fontSize: 16px;
-$CommandBarItem-chevronDown-fontSize: 12px;
-$CommandBarItem-overflow-fontSize: 18px;
-$CommandBarItem-iconLabelMargin: 8px;
-$CommandBarItem-flippedLineHeight: 32px;
-
-$SearchBox-height: 40px;
-$SearchBox-widthLgCollapsed: 50px;
-$SearchBox-widthMdDownCollapsed: 36px;
-$SearchBox-widthXLarge: 208px;
-$SearchBox-iconSize: 16px;
-$SearchBox-sidePadding: 8px; // padding in input on left and right sides without icons
+@import '~office-ui-fabric-react/src/common/common';
 
 .root {
-  @include ms-font-m;
-  color: $ms-color-neutralPrimary;
-
-  position: relative;
+  display: flex;
   background-color: $ms-color-neutralLighter;
-  height: $CommandBar-height;
-  white-space: nowrap;
-  user-select: none;
+  padding: 0 16px;
 }
 
-.container {
+.primarySet {
+  flex-grow: 1;
   display: flex;
+  align-items: stretch;
 }
 
-.primaryCommands {
-  overflow: hidden;
+.secondarySet {
+  flex-shrink: 0;
   display: flex;
-  flex: 1 1 auto;
-  vertical-align: top;
-  line-height: $CommandBar-height;
-  max-width: 100%;
-  margin: 0 20px;
+  align-items: stretch;
 }
-
-.sideCommands {
-  display: flex;
-  flex: 0 0 auto;
-  @include ms-right(0px);
-  @include text-align(right);
-  @include padding-right(20px);
-}
-
-// Command Bar Item
-.item {
-  display: inline-block;
-  color: $ms-color-themePrimary;
-  height: $CommandBarItem-height;
-  outline: transparent;
-  vertical-align: top;
-
-  &:hover {
-    background-color: $ms-color-neutralLight;
-    color: $ms-color-neutralPrimary;
-  }
-}
-
-.itemChevronDown,
-.itemCommandText {
-  display: inline-block;
-  padding: 0 4px;
-  vertical-align: top;
-}
-
-@mixin CommandBarItem-text {
-  @include focus-border(2px);
-  @include ms-font-m;
-  color: $ms-color-neutralPrimary;
-
-  position: relative;
-  background: none;
-  border: none;
-  line-height: $CommandBarItem-height;
-  min-width: 20px;
-  text-align: center;
-  padding: 0 4px;
-  display: block;
-  height: 100%;
-
-  &.itemLinkIsNoName {
-    padding: 0 8px;
-  }
-}
-
-.itemText {
-  @include CommandBarItem-text;
-}
-
-.itemLink {
-  @include CommandBarItem-text;
-  cursor: pointer;
-  &:hover,
-  &.itemLinkIsExpanded {
-    @include high-contrast {
-      @include focus-border(0px, WindowText, 1px, false);
-    }
-  }
-
-  &:hover:not([disabled]) {
-    color: $ms-color-neutralDark;
-    background-color: $ms-color-neutralLight;
-
-    .itemIcon {
-      color: $ms-color-themeDark;
-    }
-
-    .itemChevronDown {
-      color: $ms-color-neutralPrimaryAlt;
-    }
-
-    .itemOverflow {
-      color: $ms-color-neutralDark;
-    }
-  }
-
-  &.itemLinkIsExpanded {
-    background-color: $ms-color-neutralQuaternaryAlt;
-    color: $ms-color-black;
-
-    .itemIcon {
-      color: $ms-color-themeDarker;
-    }
-
-    .itemChevronDown {
-      color: $ms-color-neutralDark;
-    }
-
-    .itemOverflow {
-      color: $ms-color-black;
-    }
-  }
-
-  &.itemLinkIsExpanded:hover {
-    background-color: $ms-color-neutralQuaternary;
-  }
-
-  &[disabled] {
-    color: $ms-color-neutralTertiaryAlt;
-    cursor: default;
-    pointer-events: none;
-    .itemIcon {
-      color: $ms-color-neutralTertiaryAlt;
-    }
-  }
-}
-
-.itemIcon {
-  font-size: $CommandBarItem-fontSize;
-  padding: 0 4px;
-}
-
-.itemIconColor {
-  color: $ms-color-themeDarkAlt;
-}
-
-.itemChevronDown  {
-  color: $ms-color-neutralSecondary;
-  font-size: $CommandBarItem-chevronDown-fontSize;
-}
-
-.itemOverflow {
-  font-size: $CommandBarItem-overflow-fontSize;
-  color: $ms-color-neutralPrimary;
-  padding: 0 7px;
-}
-
-// CommandBarSearch needs to replaced with SearchBox component, and hence following styles revisited/cleaned.
-.search {
-  @include float(left);
-  width: 208px;
-  max-width: 208px;
-  background-color: $ms-color-white;
-  color: $ms-color-neutralPrimary;
-  height: $SearchBox-height;
-  position: relative;
-  box-sizing: border-box;
-  border-color: transparent; // give a base state for animation
-  @include border-right(1px, solid, $ms-color-neutralLight);
-
-  @include high-contrast {
-    @include border-right(1px, solid, WindowText);
-    z-index: 10;
-  }
-
-  &:before {
-    position: absolute;
-    content: ' ';
-    right: 0;
-    bottom: 0;
-    left: 0;
-    margin: 0 8px;
-    border-bottom: 1px solid $ms-color-neutralLight;
-  }
-
-  &:hover {
-    color: $ms-color-themePrimary;
-
-    @include high-contrast {
-      @include focus-border(0px, WindowText, 1px, false);
-    }
-  }
-}
-
-.searchInput {
-  height: $SearchBox-height;
-  // Due to a bug in IE9, we have to use a transparent left border instead of left padding
-  padding: $SearchBox-sidePadding $SearchBox-sidePadding $SearchBox-sidePadding 0;
-  border: none;
-  @include border-left($SearchBox-height + 2, solid, transparent);
-  background-color: transparent;
-  width: 100%;
-  box-sizing: border-box;
-  outline: none;
-  cursor: pointer;
-  font-size: $ms-font-size-m;
-
-  @include high-contrast {
-    @include border-left($SearchBox-height, solid, Window);
-  }
-
-  // remove default iOS styles on input boxes
-  -webkit-appearance: none;
-  -webkit-border-radius: 0;
-
-  // hide IE's X button for clearing the input (we have our own)
-  &::-ms-clear {
-    display: none;
-  }
-}
-
-// default styles for icons, show the search icon and hide X (and arrow) until focused
-.searchIconSearchWrapper {
-  display: block;
-}
-
-.searchIconArrowWrapper {
-  display: none;
-}
-
-.searchIconSearchWrapper,
-.searchIconArrowWrapper {
-  top: 0;
-  @include ms-padding-left(17px);
-  @include padding-right(8px);
-}
-
-.searchIconClearWrapper {
-  display: none;
-  top: 1px;
-  @include ms-right(0px);
-  z-index: $ms-zIndex-front;
-}
-
-.searchIconWrapper {
-  height: $SearchBox-height;
-  line-height: $SearchBox-height;
-  cursor: pointer;
-  position: absolute;
-  text-align: center;
-  -ms-high-contrast-adjust: none;
-}
-
-// New rules
-.search :global(.ms-Icon:before) {
-  font-size: $SearchBox-iconSize;
-  color: $ms-color-themePrimary;
-}
-
-.searchInput,
-.searchInput::-webkit-input-placeholder {
-  font-size: $ms-font-size-m;
-}
-
