+++ packages/office-ui-fabric-react/src/components/CommandBar/examples/CommandBar.Basic.Example.tsx
@@ -1,71 +1,27 @@
 import * as React from 'react';
-import { assign } from 'office-ui-fabric-react/lib/Utilities';
-import { CommandBar } from 'office-ui-fabric-react/lib/CommandBar';
-import { Toggle } from 'office-ui-fabric-react/lib/Toggle';
+import { CommandBar } from '../CommandBar';
+import { ICommandBarProps } from './../CommandBar.Props';
 
-export class CommandBarBasicExample extends React.Component<any, any> {
+export class CommandBarBasicExample extends React.Component<ICommandBarProps, {}> {
 
-  constructor(props: any) {
+  constructor(props: ICommandBarProps) {
     super(props);
     this.state = {
-      isSearchBoxVisible: true,
       areNamesVisible: true,
       areIconsVisible: true
     };
   }
 
-  public render() {
+  public render(): JSX.Element {
     let { items, overflowItems, farItems } = this.props;
-    let { isSearchBoxVisible: searchBoxVisible, areIconsVisible: iconsVisible, areNamesVisible: namesVisible } = this.state;
-
-    let filteredItems = items.map((item: any) => assign({}, item, {
-      name: namesVisible ? item.name : '',
-      icon: iconsVisible ? item.icon : ''
-    }));
-
-    let filteredOverflowItems = overflowItems.map((item: any) => assign({}, item, {
-      name: namesVisible ? item.name : '',
-      icon: iconsVisible ? item.icon : ''
-    }));
-
-    let filteredFarItems = farItems.map((item: any) => assign({}, item, {
-      name: namesVisible ? item.name : '',
-      icon: iconsVisible ? item.icon : ''
-    }));
 
     return (
       <div>
-        <Toggle
-          label='Show search box'
-          checked={ searchBoxVisible }
-          // tslint:disable-next-line:jsx-no-lambda
-          onChanged={ isSearchBoxVisible => this.setState({ isSearchBoxVisible }) }
-          onText='Visible'
-          offText='Hidden'
-        />
-        <Toggle
-          label='Show names'
-          checked={ namesVisible }
-          // tslint:disable-next-line:jsx-no-lambda
-          onChanged={ areNamesVisible => this.setState({ areNamesVisible }) }
-          onText='Visible'
-          offText='Hidden'
-        />
-        <Toggle
-          label='Show icons'
-          checked={ iconsVisible }
-          // tslint:disable-next-line:jsx-no-lambda
-          onChanged={ areIconsVisible => this.setState({ areIconsVisible }) }
-          onText='Visible'
-          offText='Hidden'
-        />
         <CommandBar
-          isSearchBoxVisible={ searchBoxVisible }
-          searchPlaceholderText='Search...'
           elipisisAriaLabel='More options'
-          items={ filteredItems }
-          overflowItems={ filteredOverflowItems }
-          farItems={ filteredFarItems }
+          items={ items }
+          overflowItems={ overflowItems }
+          farItems={ farItems }
         />
       </div>
     );
