--- packages/office-ui-fabric-react/src/components/CommandBar/examples/CommandBar.Example.scss
@@ -1,48 +1,0 @@
-@import '../../../common/common';
-
-.customButtonContainer {
-    display: flex;
-    align-items: center;
-}
-
-.icon {
-    @include ms-Icon;
-    font-size: 14px;
-    position: relative;
-}
-
-.themeDarkAltColor {
-    color: $ms-color-themeDarkAlt;
-}
-
-.leftText {
-    margin-left: 6px;
-    @include ms-font-m;
-    color: $ms-color-neutralPrimary;
-}
-
-.splitter {
-    @include ms-font-m;
-    color: $ms-color-neutralTertiary;
-    background: transparent;
-    position: relative;
-    top: -2px;
-    height: 40px;
-}
-
-/* Require update when office-ui-fabric-react changes to CSS modules */
-.button {
-    &:global.ms-Button {
-        height: 40px;
-        min-width: 28px;
-        padding: 0 5px;
-    }
-
-    &:hover {
-      background: initial;
-    }
-}
-
-.darkerBG {
-  background: #d0d0d0
-}
