--- packages/office-ui-fabric-react/src/components/CommandBar/examples/CommandBar.NonFocusable.Example.tsx
@@ -1,18 +1,0 @@
-import * as React from 'react';
-import { CommandBar } from 'office-ui-fabric-react/lib/CommandBar';
-import { itemsNonFocusable, farItemsNonFocusable } from './data-nonFocusable';
-
-export class CommandBarNonFocusableItemsExample extends React.Component<any, any> {
-  public render() {
-    return (
-      <div>
-        <CommandBar
-          isSearchBoxVisible={ false }
-          items={ itemsNonFocusable }
-          farItems={ farItemsNonFocusable }
-        />
-      </div>
-    );
-  }
-
-}
