+++ packages/office-ui-fabric-react/src/components/CommandBar/examples/data.ts
@@ -2,7 +2,10 @@ export const items = [
   {
     key: 'newItem',
     name: 'New',
-    icon: 'Add',
+    cacheKey: 'myCacheKey',
+    iconProps: {
+      iconName: 'Add',
+    },
     ariaLabel: 'New. Use left and right arrow keys to navigate',
     ['data-automation-id']: 'newItemMenu',
     subMenuProps: {
@@ -10,13 +13,17 @@ export const items = [
         {
           key: 'emailMessage',
           name: 'Email message',
-          icon: 'Mail',
+          iconProps: {
+            iconName: 'Mail',
+          },
           ['data-automation-id']: 'newEmailButton'
         },
         {
           key: 'calendarEvent',
           name: 'Calendar event',
-          icon: 'Calendar'
+          iconProps: {
+            iconName: 'Calendar'
+          },
         }
       ],
     },
@@ -24,90 +31,26 @@ export const items = [
   {
     key: 'upload',
     name: 'Upload',
-    icon: 'Upload',
-    href: 'https://mytenenat.sharepoint.com/teams/IT/BPU/',
+    iconProps: {
+      iconName: 'Upload',
+    },
+    href: 'https://dev.office.com/fabric',
     ['data-automation-id']: 'uploadButton'
   },
   {
     key: 'share',
     name: 'Share',
-    icon: 'Share',
+    iconProps: {
+      iconName: 'Share',
+    },
     onClick: () => { return; }
   },
   {
     key: 'download',
     name: 'Download',
-    icon: 'Download',
-    onClick: () => { return; }
-  },
-  {
-    key: 'disabled',
-    name: 'Disabled...',
-    icon: 'Cancel',
-    disabled: true,
-    onClick: () => { return; }
-  }
-];
-
-export const textOnlyItems = [
-  {
-    key: 'upload',
-    name: 'Upload',
-    onClick: () => { return; }
-  },
-  {
-    key: 'share',
-    name: 'Share',
-    onClick: () => { return; }
-  },
-  {
-    key: 'download',
-    name: 'Download',
-    onClick: () => { return; }
-  }
-];
-
-export const iconOnlyItems = [
-  {
-    key: 'upload',
-    name: '',
-    icon: 'Upload',
-    onClick: () => { return; }
-  },
-  {
-    key: 'share',
-    name: '',
-    icon: 'Share',
-    onClick: () => { return; }
-  },
-  {
-    key: 'download',
-    name: '',
-    icon: 'Download',
-    onClick: () => { return; }
-  },
-  {
-    key: 'move',
-    name: '',
-    icon: 'MoveToFolder',
-    onClick: () => { return; }
-  },
-  {
-    key: 'copy',
-    name: '',
-    icon: 'Copy',
-    onClick: () => { return; }
-  },
-  {
-    key: 'rename',
-    name: '',
-    icon: 'Edit',
-    onClick: () => { return; }
-  },
-  {
-    key: 'disabled',
-    icon: 'Cancel',
-    disabled: true,
+    iconProps: {
+      iconName: 'Download',
+    },
     onClick: () => { return; }
   }
 ];
@@ -116,17 +59,23 @@ export const overflowItems = [
   {
     key: 'move',
     name: 'Move to...',
-    icon: 'MoveToFolder'
+    iconProps: {
+      iconName: 'MoveToFolder'
+    },
   },
   {
     key: 'copy',
     name: 'Copy to...',
-    icon: 'Copy'
+    iconProps: {
+      iconName: 'Copy'
+    },
   },
   {
     key: 'rename',
     name: 'Rename...',
-    icon: 'Edit'
+    iconProps: {
+      iconName: 'Edit'
+    },
   }
 ];
 
@@ -134,19 +83,27 @@ export const farItems = [
   {
     key: 'sort',
     name: 'Sort',
-    icon: 'SortLines',
+    iconProps: {
+      iconName: 'SortLines',
+    },
     onClick: () => { return; }
   },
   {
     key: 'tile',
     name: 'Grid view',
-    icon: 'Tiles',
+    iconProps: {
+      iconName: 'Tiles',
+    },
+    iconOnly: true,
     onClick: () => { return; }
   },
   {
     key: 'info',
     name: 'Info',
-    icon: 'Info',
+    iconProps: {
+      iconName: 'Info',
+    },
+    iconOnly: true,
     onClick: () => { return; }
   }
 ];
